Create dedicated namespaces for Admin and Api listeners and update the service definitions
