+++ app/config/oauth-api/services.yml
@@ -13,7 +13,7 @@ services:
   # CRITICAL this listener must have a priority greater than ApiPlatform ReadListener (which has a priority of 4)
   # but it must be run AFTER the Firewall listener or the Security token won't be initialized yet (which has a
   # priority of 8)
-  PrestaShopBundle\EventListener\ScopeCheckerListener:
+  PrestaShopBundle\EventListener\API\ScopeCheckerListener:
     autowire: true
     tags:
       - { name: kernel.event_listener, event: kernel.request, method: onKernelRequest, priority: 7 }
