+++ src/PrestaShopBundle/EventListener/API/ScopeCheckerListener.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShopBundle\EventListener;
+namespace PrestaShopBundle\EventListener\API;
 
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
 use ApiPlatform\State\Util\OperationRequestInitiatorTrait;
