+++ tests/Integration/PrestaShopBundle/EventListener/Admin/AccessDeniedListenerTest.php
@@ -24,7 +24,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  */
 
-namespace Tests\Integration\PrestaShopBundle\EventListener;
+namespace Tests\Integration\PrestaShopBundle\EventListener\Admin;
 
 use Symfony\Bundle\FrameworkBundle\KernelBrowser;
 use Symfony\Bundle\FrameworkBundle\Routing\Router;
